backend/pyproject.toml
[project]
name = "shadowtrader-backend"
version = "0.1.0"
description = "ShadowTrader AI backend"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.13.1",
  "alpaca-py>=0.28.1",
  "anthropic>=0.46.0",
  "asyncpg>=0.29.0",
  "celery[redis]>=5.3.6",
  "cryptography>=42.0.5",
  "fastapi>=0.110.0",
  "greenlet>=3.0.3",
  "gunicorn>=23.0.0",
  "httpx>=0.27.0",
  "ib_insync>=0.9.86",
  "pydantic-settings>=2.2.1",
  "python-binance>=1.0.19",
  "python-dotenv>=1.0.1",
  "python-jose[cryptography]>=3.3.0",
  "redis>=5.0.3",
  "sqlalchemy>=2.0.29",
  "stripe>=11.6.0",
  "supabase>=2.11.0",
  "uvicorn[standard]>=0.29.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.1.1",
  "pytest-asyncio>=0.23.6",
  "ruff>=0.3.4",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "C4"]
